version 0005: 更新 Exercise05，加上 PrettyTable 與相關設定

+++ Student_dataClass.py
@@ -6,10 +6,10 @@ class Student:
     first_name: str
     second_name: str
     phone_number: str
-    grade_list: list[list[int]] = field(init=False)
+    grade_list: list[list[float]] = field(init=False)
 
     def __post_init__(self):
         self.grade_list = []
 
-    def add_grade(self, grade: list[int]) -> None:
+    def add_grade(self, grade: list[float]) -> None:
         self.grade_list.append(grade)
